Convert fluxscp to a proper flux sink.
