+++ lib/fluxsink/fluxsink.proto
@@ -25,12 +25,22 @@ message VcdFluxSinkProto {
 		(help) = "directory to write .vcd files to"];
 }
 
+message ScpFluxSinkProto {
+	optional string filename = 2 [default = "flux.scp",
+		(help) = ".scp file to write to"];
+	optional bool align_with_index = 3 [default = false,
+		(help) = "align data to track boundaries"];
+	optional int32 type_byte = 4 [default = 0xff,
+		(help) = "set the SCP disk type byte"];
+}
+
 message FluxSinkProto {
 	oneof dest {
 		string fluxfile = 1;
 		HardwareFluxSinkProto drive = 2;
 		AuFluxSinkProto au = 3;
 		VcdFluxSinkProto vcd = 4;
+		ScpFluxSinkProto scp = 5;
 	}
 }
