+++ src/fe-rawread.cc
@@ -65,6 +65,15 @@ static StringFlag vcdFile(
 		config.mutable_output()->mutable_flux()->mutable_vcd()->set_directory(value);
 	});
 
+static StringFlag scpFile(
+	{ "--scp" },
+	"write destination flux to this .scp file",
+	"",
+	[](const auto& value)
+	{
+		config.mutable_output()->mutable_flux()->mutable_scp()->set_filename(value);
+	});
+
 static StringFlag srcCylinders(
 	{ "--cylinders", "-c" },
 	"cylinders to read from",
